LIG-9843: refactor evaluation routes into a barrel package
Move the existing GET routes (runs, sample metrics info, confusion matrix) from
evaluation.py into an evaluation/ package, each with its own APIRouter, and a
barrel __init__ that aggregates them. Update the test's mock paths to the new
module locations. Pure refactor; no endpoint or behavior change.
